refactor doozer rpms: commands to rpms.py
